refactor(enum): retro - type CoAP public request/response fields
CoapRequest.method -> CoapMethod, CoapRequest/CoapResponse.content_format ->
CoapContentFormat, and emit_header/emit_options_payload + the local type/rsp_type/
req_cf take the enums, so a handler switches on req->method and sets
content_format enum-to-enum (a wrong value no longer compiles). CoapResponse.code
stays uint8_t - it is the bit-packed CoAP code octet (class<<5|detail, read via
>>5), a genuine wire byte; the (uint8_t)CoapResponseCode casts into it and the
CoapMethodMask bitmask ORs are wire/flags boundaries, kept. Examples + READMEs
updated to the scoped API. native_coap + native_coap_observe green (90).

+++ examples/L7-Application/13.CoAP/13.CoAP.ino
@@ -56,14 +56,14 @@ static void coap_info(const CoapRequest *req, CoapResponse *resp)
     if (n < 0)
         n = 0;
     resp->payload_len = (size_t)n;
-    resp->content_format = (uint16_t)CoapContentFormat::COAP_CF_JSON;
+    resp->content_format = CoapContentFormat::COAP_CF_JSON;
     resp->code = (uint8_t)CoapResponseCode::COAP_RSP_CONTENT;
 }
 
 // GET/PUT /led -> read or drive the on-board LED.
 static void coap_led(const CoapRequest *req, CoapResponse *resp)
 {
-    if (req->method == (uint8_t)CoapMethod::COAP_PUT)
+    if (req->method == CoapMethod::COAP_PUT)
     {
         // Treat any payload beginning with a non-'0' character as "on".
         g_led_state = (req->payload_len && req->payload[0] != '0') ? 1 : 0;
@@ -74,7 +74,7 @@ static void coap_led(const CoapRequest *req, CoapResponse *resp)
     }
     resp->payload[0] = g_led_state ? '1' : '0';
     resp->payload_len = 1;
-    resp->content_format = (uint16_t)CoapContentFormat::COAP_CF_TEXT;
+    resp->content_format = CoapContentFormat::COAP_CF_TEXT;
     resp->code = (uint8_t)CoapResponseCode::COAP_RSP_CONTENT;
 }
 
@@ -88,7 +88,7 @@ static void coap_hello(const CoapRequest *req, CoapResponse *resp)
         n = resp->payload_cap;
     memcpy(resp->payload, msg, n);
     resp->payload_len = n;
-    resp->content_format = (uint16_t)CoapContentFormat::COAP_CF_TEXT;
+    resp->content_format = CoapContentFormat::COAP_CF_TEXT;
     resp->code = (uint8_t)CoapResponseCode::COAP_RSP_CONTENT;
 }
